preserve existing .env when running docker_dev_setup
flag=none

Test Plan:
1. Create .env file with custom COMPOSE_FILE overrides.
2. Run script/docker_dev_setup.sh
3. Verify custom compose overrides are still configured.

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/259027
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Bryan Petty <bpetty@instructure.com>
Product-Review: Bryan Petty <bpetty@instructure.com>

+++ script/docker_dev_setup.sh
@@ -208,6 +208,12 @@ function setup_docker_environment {
     message "Copying default configuration from config/docker-compose.override.yml.example to docker-compose.override.yml"
     cp config/docker-compose.override.yml.example docker-compose.override.yml
   fi
+
+  if [ -f ".env" ]; then
+    prompt '.env file exists, would you like to reset it to default? [y/n]' confirm
+    [[ ${confirm:-n} == 'y' ]] || return 0
+  fi
+  message "Setting up default .env configuration"
   echo -n "COMPOSE_FILE=docker-compose.yml:docker-compose.override.yml" > .env
 }
